Edit RestCard styles and refactor shared/ui
