+++ src/entities/Dish/ui/DishCard/styles.module.scss
@@ -2,6 +2,14 @@
 @use '@app/styles/mixins' as m;
 
 .dish_card {
+    display: flex;
+    flex-direction: column;
+    align-items: start;
+    justify-content: start;
+    width: 100%;
+    height: 100%;
+    padding: 20px;
+
     &__image {
         width: 184px;
 
@@ -11,6 +19,7 @@
             align-items: center;
             justify-content: center;
             // width: 184px;
+            // padding: 20px 20px 0px 20px;
             user-select: none;
             cursor: default;
         }
