+++ src/entities/Restaurant/ui/RestaurantCard/styles.module.scss
@@ -2,8 +2,17 @@
 @use '@app/styles/mixins' as m;
 
 .restaurant_card {
+    &__container {
+        display: flex;
+        flex-direction: column;
+        align-items: start;
+        justify-content: start;
+        width: 100%;
+        height: 100%;
+        padding: 20px;
+    }
     &__image {
-        width: 184px;
+        width: 100%;
 
         &_container {
             display: flex;
@@ -36,6 +45,14 @@
         flex-direction: column;
         align-items: start;
         justify-content: start;
+        width: 100%;
+    }
+
+    &__features_container {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        justify-content: space-between;
         width: 100%;
     }
 }
