+++ src/features/like-product/ui/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import styles from './styles.module.scss';
-import { HeartSvg } from '@shared/ui/svg-components/HeartSvg';
+import { HeartSvg } from '@shared/ui/icon-components/HeartSvg';
 
 export const LikeProduct = () => {
     const [like, setLike] = useState(false);
