+++ src/shared/ui/buttons/view-all-button/index.tsx
