+++ src/shared/ui/buttons/view-all-button/styles.module.scss
