+++ src/shared/ui/elements/card-wrapper/index.tsx
@@ -1,6 +1,10 @@
 import styles from './styles.module.scss';
 import { ICardWrapper } from './types';
 
-export const CardWrapper: React.FC<ICardWrapper> = ({ children }) => {
-    return <div className={styles.card_wrapper}>{children}</div>;
+export const CardWrapper: React.FC<ICardWrapper> = ({ children, width }) => {
+    return (
+        <div className={styles.card_wrapper} style={{ width: width }}>
+            {children}
+        </div>
+    );
 };
