+++ src/shared/ui/elements/card-wrapper/styles.module.scss
@@ -9,10 +9,8 @@
     list-style-type: none;
     background-color: v.$white;
     border-radius: 34px;
-    width: 224.63px;
-    height: 390px;
-    // height: 100%;
-    padding: 20px;
+    min-width: 140px;
+    min-height: 230px;
     box-shadow: 5px 5px 15px rgba(0, 0, 0, 0.15);
     transition: transform 0.3s ease;
     transform-origin: center;
