+++ src/shared/ui/elements/card-wrapper/types.ts
@@ -2,4 +2,5 @@ import { ReactNode } from 'react';
 
 export interface ICardWrapper {
     children: ReactNode;
+    width?: string;
 }
