+++ src/shared/ui/elements/delivery-and-raiting/index.tsx
