+++ src/shared/ui/elements/index.ts
@@ -1,2 +1,6 @@
 export { CategoryElement } from './category-element';
 export { CardWrapper } from './card-wrapper';
+export { DeliveryAndRaiting } from './delivery-and-raiting';
+export { PageContainer } from './page-container';
+export { LogoLink } from './logo-link';
+export { SeparatorLine } from './separator-line';
