+++ src/shared/ui/elements/page-container/styles.module.scss
