+++ src/shared/ui/icon-components/HeartSvg.tsx
