+++ src/shared/ui/icon-components/ProfileSVG.tsx
