+++ src/shared/ui/icon-components/custom-icon/index.tsx
