+++ src/shared/ui/icon-components/custom-icon/styles.module.scss
