+++ src/shared/ui/icon-components/custom-icon/types.ts
