+++ src/shared/ui/icon-components/index.ts
@@ -1,3 +1,4 @@
 export { HeartSvg } from './HeartSvg';
 export { ProfileSVG } from './ProfileSVG';
 export { ArrowSvg } from './ArrowSvg';
+export { CustomIcon } from './custom-icon/';
