+++ src/shared/ui/index.ts
@@ -1,9 +1,11 @@
-export { PageContainer } from './page-container/index';
-export { ReusableButton } from './reusable-button/index';
-export { LogoLink } from './logo-link/index';
-export { CustomIcon } from './custom-icon/index';
-export { TopListSection } from './top-list-section/index';
-export { SeparatorLine } from './separator-line/index';
-export { HeartSvg, ProfileSVG, ArrowSvg } from './svg-components';
-export { CardWrapper, CategoryElement } from './elements/';
-export { DeliveryAndRaiting } from './delivery-and-raiting';
+export {
+    CardWrapper,
+    CategoryElement,
+    PageContainer,
+    LogoLink,
+    SeparatorLine,
+    DeliveryAndRaiting,
+} from './elements/';
+export { ReusableButton, AddButton, ViewAllButton } from './buttons/';
+export { HeartSvg, ProfileSVG, ArrowSvg, CustomIcon } from './icon-components';
+export { TopListSection } from './templates/';
