+++ src/shared/ui/templates/index.ts
@@ -1,0 +1,1 @@
+export { TopListSection } from './top-list-section/index';
