+++ src/shared/ui/templates/top-list-section/styles.module.scss
