+++ src/shared/ui/templates/top-list-section/types.ts
